feat: implement access token authentication middleware
